Move pkg/apps/appfs to pkg/appfs
