+++ pkg/apps/webapp.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/cozy/afero"
 	"github.com/cozy/cozy-stack/model/notification"
-	"github.com/cozy/cozy-stack/pkg/apps/appfs"
+	"github.com/cozy/cozy-stack/pkg/appfs"
 	"github.com/cozy/cozy-stack/pkg/consts"
 	"github.com/cozy/cozy-stack/pkg/couchdb"
 	"github.com/cozy/cozy-stack/pkg/jobs"
